split test directory into terst & src
